[BUGFIX] Prevent exception when accessing Install Tool from Backend
BEWARE: The dialog "Confirm with user password" to access the various
modules of the Install Tool does not support 2FA so trying to confirm
with the password of someone having 2FA enabled will always fail at
this point. It will just not crash anymore.

The corollary is that for such cases, only the Install Tool password
may be used.