[TASK] Drop extension configuration for enabling 2FA in Frontend
Since TYPO3 v11, 2FA in Backend is natively handled. As such corresponding
code has been dropped and existing 2FA configuration for be_users is
migrated to new native field "mfa".

This means that when someone uses this extension in TYPO3 v11, they logically
want 2FA for the Frontend; and thus there is no need for a configuration
flag anymore.